Renamed auth.py into handlers.py
